split store into multiple files

+++ src/components/SessionDetail.vue
@@ -1,20 +1,49 @@
 <template>
   <div>
-    <div class="title">
-      {{ item.title }}
-      <v-icon
-        class="favorite"
-        @click="toggleFavorite(item)"
-      >{{ item.isFavorite ? 'turned_in' : 'turned_in_not' }}</v-icon>
+    <div class="back">
+      <v-btn flat small color="primary" @click="back">Back to Sessions</v-btn>
     </div>
-    <span class="speaker">{{ item.speaker }}</span>
-    <div class="subheading">{{ item.description }}</div>
-    <div class="tags">
-      <v-chip v-for="tag in item.tags" :key="tag">{{ tag }}</v-chip>
+    <div class="detail-wrapper">
+      <div class="title">
+        {{ item.title }}
+        <v-icon
+          class="favorite"
+          @click="toggleFavorite(item)"
+        >{{ item.isFavorite ? 'turned_in' : 'turned_in_not' }}</v-icon>
+        <div class="delete">
+          <v-menu bottom right>
+            <v-btn slot="activator" icon>
+              <v-icon>more_vert</v-icon>
+            </v-btn>
+            <v-list>
+              <v-list-tile>
+                <v-list-tile-action>
+                  <v-icon>delete</v-icon>
+                </v-list-tile-action>
+                <v-list-tile-title class="clickable" @click="deleteSession(item.id)">Delete Session</v-list-tile-title>
+              </v-list-tile>
+              <v-list-tile>
+                <v-list-tile-action>
+                  <v-icon>edit</v-icon>
+                </v-list-tile-action>
+                <v-list-tile-title class="clickable" @click="editSession(item.id)">Edit Session</v-list-tile-title>
+              </v-list-tile>
+            </v-list>
+          </v-menu>
+        </div>
+      </div>
+      <span class="speaker">{{ item.speaker }}</span>
+      <div class="subheading">{{ item.description }}</div>
+      <div class="tags">
+        <v-chip v-for="tag in item.tags" :key="tag">{{ tag }}</v-chip>
+      </div>
     </div>
-    <v-card-actions>
-      <v-btn @click="deleteSession(item.id)">Delete</v-btn>
-    </v-card-actions>
+    <ConfirmModal
+      ref="deleteModal"
+      prompt="Delete session?"
+      agreeText="Delete"
+      disagreeText="Cancel"
+    />
   </div>
 </template>
 
@@ -22,8 +51,13 @@
 import { Component, Prop, Vue } from 'vue-property-decorator';
 import { AppState, Session } from '@/types';
 import { State, Getter, Mutation, Action } from 'vuex-class';
+import ConfirmModal from '@/components/ConfirmModal.vue';
 
-@Component({})
+@Component({
+  components: {
+    ConfirmModal
+  }
+})
 export default class SessionDetail extends Vue {
   public item!: Session;
 
@@ -46,19 +80,35 @@ export default class SessionDetail extends Vue {
     super();
   }
 
-  public deleteSession(id: string): void {
-    this.deleteSessionAsync(id);
-    this.$router.replace('/sessions');
-    this.showAlert('Session deleted');
+  protected created() {
+    this.item = this.getSessionById(this.id);
   }
 
-  public created() {
-    this.item = this.getSessionById(this.id);
+  private deleteSession(id: string): void {
+    const deleteModal = this.$refs.deleteModal as ConfirmModal;
+    deleteModal.show().then(confirmAction => {
+      if (confirmAction) {
+        this.deleteSessionAsync(id);
+        this.$router.replace('/sessions');
+        this.showAlert('Session deleted');
+      }
+    });
+  }
+
+  private back() {
+    this.$router.replace('/sessions');
   }
 }
 </script>
 
 <style>
+.detail-wrapper {
+  padding-left: 15px;
+}
+
+.back {
+  padding-bottom: 15px;
+}
 .favorite {
   cursor: pointer;
   padding-left: 10px;
@@ -70,5 +120,9 @@ export default class SessionDetail extends Vue {
 
 .tags {
   padding-top: 15px;
+}
+
+.delete {
+  float: right;
 }
 </style>
